# Secp256r1 precompile: switch to BoringSSL