--- a/src/Nethermind/Nethermind.Evm/Precompiles/Secp256r1Precompile.cs
+++ b/src/Nethermind/Nethermind.Evm/Precompiles/Secp256r1Precompile.cs
@@ -2,10 +2,10 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System;
-using System.Security.Cryptography;
 using Nethermind.Core;
 using Nethermind.Core.Extensions;
 using Nethermind.Core.Specs;
+using Nethermind.Crypto;
 
 namespace Nethermind.Evm.Precompiles;
 
@@ -19,37 +19,13 @@ public class Secp256r1Precompile : IPrecompile<Secp256r1Precompile>
     public long BaseGasCost(IReleaseSpec releaseSpec) => 3450L;
     public long DataGasCost(ReadOnlyMemory<byte> inputData, IReleaseSpec releaseSpec) => 0L;
 
-    // TODO can be optimized - Go implementation is 2-6 times faster depending on the platform. Options:
-    // - Try to replicate Go version in C#
-    // - Compile Go code into a library and call it via P/Invoke
     public (byte[], bool) Run(ReadOnlyMemory<byte> inputData, IReleaseSpec releaseSpec)
     {
-        if (inputData.Length != 160)
-            return (null, true);
-
-        ReadOnlySpan<byte> bytes = inputData.Span;
-        ReadOnlySpan<byte> hash = bytes[..32], sig = bytes[32..96];
-        ReadOnlySpan<byte> x = bytes[96..128], y = bytes[128..160];
-
-        ECDsa ecdsa;
-        try
-        {
-            ecdsa = ECDsa.Create(new ECParameters
-            {
-                Curve = ECCurve.NamedCurves.nistP256,
-                Q = new() { X = x.ToArray(), Y = y.ToArray() }
-            });
-        }
-        catch
-        {
-            // Invalid x/y parameters
-            return (null, true);
-        }
-
-        var isValid = ecdsa.VerifyHash(hash, sig);
-
         Metrics.Secp256r1Precompile++;
 
-        return (isValid ? ValidResult : null, true);
+        return (
+            Secp256r1.VerifySignature(inputData) ? ValidResult : null,
+            true
+        );
     }
 }