## Changes

- Switches to the [secp256r1 package](https://github.com/NethermindEth/secp256r1-bindings) using BoringSSL for signature verification (see https://github.com/NethermindEth/secp256r1-bindings/pull/6)
- Adds a bunch of random test cases
- Small fixes for Benchmarks solution:
  - Broken build because of missing projects
  - Precompile benchmarks runnable only from specific directory
  - 2 baseline methods in single benchmark

Related:
- https://github.com/NethermindEth/nethermind/pull/8252

## Types of changes

#### What types of changes does your code introduce?

- [x] Optimization

## Testing

#### Requires testing

- [x] Yes

#### If yes, did you write tests?

- [x] Yes

## Documentation

#### Requires documentation update

- [x] No

#### Requires explanation in Release Notes

- [x] No