@alexb5dh 2025-03-28
### Windows 10


| Type               | Method   | Precompile Name     | Input    | Mean     | Error    | StdDev   | Ratio | RatioSD | Gas  | Throughput   | Throughput CI-Lower | Throughput CI-Upper | Allocated | Alloc Ratio |
|------------------- |--------- |-------------------- |--------- |---------:|---------:|---------:|------:|--------:|-----:|-------------:|--------------------:|--------------------:|----------:|------------:|
| Secp256r1Benchmark | Baseline | Secp256r1Precompile | Invalid  | 53.20 us | 0.091 us | 0.130 us |  1.00 |    0.00 | 3450 | 64.85 MGas/s |        64.93 MGas/s |        64.76 MGas/s |         - |          NA |
|                    |          |                     |          |          |          |          |       |         |      |              |                     |                     |           |             |
| Secp256r1Benchmark | Baseline | Secp256r1Precompile | Valid1   | 53.28 us | 0.104 us | 0.153 us |  1.00 |    0.00 | 3450 | 64.75 MGas/s |        64.85 MGas/s |        64.66 MGas/s |         - |          NA |
|                    |          |                     |          |          |          |          |       |         |      |              |                     |                     |           |             |
| Secp256r1Benchmark | Baseline | Secp256r1Precompile | Valid2   | 53.29 us | 0.163 us | 0.239 us |  1.00 |    0.01 | 3450 | 64.74 MGas/s |        64.89 MGas/s |        64.59 MGas/s |         - |          NA |
|                    |          |                     |          |          |          |          |       |         |      |              |                     |                     |           |             |
| Secp256r1Benchmark | Baseline | Secp256r1Precompile | Valid3   | 53.68 us | 0.123 us | 0.177 us |  1.00 |    0.00 | 3450 | 64.27 MGas/s |        64.38 MGas/s |        64.16 MGas/s |         - |          NA |
|                    |          |                     |          |          |          |          |       |         |      |              |                     |                     |           |             |
| EcRecoverBenchmark | Baseline | EcRecoverPrecompile | ValidKey | 57.73 us | 1.063 us | 1.559 us |  1.00 |    0.04 | 3000 | 51.96 MGas/s |        52.69 MGas/s |        51.25 MGas/s |     416 B |        1.00 |

### Ubuntu 22 (WSL)

**WIP**

*Allocated doesn't take unmanaged memory into account